cram-school-attendance/requirements.txt
gspread
google-auth
qrcode
flask